// src/app/core/models/water-quality.model.ts
export interface testing_points {
  id: string,
  organizationId:string,
  pointCode: string,
  pointName: string,
  pointType: PointType, 
  zoneId: string,
  streetId?: string, // Optional field for DOMICILIO type points
  locationDescription: string,
  coordinates: {
    latitude: number,
    longitude: number
  },
  status: Status,
  created_at: string
}

export enum PointType{
  RESERVORIO ='RESERVORIO',
  RED_DISTRIBUCION='RED_DISTRIBUCION',
  DOMICILIO='DOMICILIO'
}
// ...
export interface QualityTest {
  id:string;
  organizationId: string;
  testCode: string;
  testingPointId: string;
  testDate: string; 
  testType: TestType;
  testedByUserId: string;
  weatherConditions: string;
  waterTemperature: number;
  generalObservations: string;
  status: 'COMPLETED'; // podría también aceptar otros estados si aplican
  results: TestResult[];
}
// ...
export interface TestResult {
  parameterId?: string; // opcional si no se está enviando
  parameterCode: string;
  measuredValue: number;
  unit: string;
  status: StatusResult;
  observations: string;
}

export enum StatusResult{
  ACCEPTABLE = 'ACCEPTABLE',
  WARNING = 'WARNING',
  CRITICAL = 'CRITICAL'
}

// ENUM TESTTYPE
export enum TestType{
  RUTINARIO='RUTINARIO',
  ESPECIAL='ESPECIAL',
  INCIDENCIA='INCIDENCIA'
}
// ...
export enum Status{
    ACTIVE ='ACTIVE',
    INACTIVE = 'INACTIVE'
}